feat: fix header logic,style, add new route for profile-edit

+++ src/app/protected/dashboard/dashboard.component.ts
@@ -1,5 +1,7 @@
 import { Router } from '@angular/router';
 import { Component } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmModalComponent } from '../components/confirm-modal-component/confirm-modal.component';
 
 @Component({
   selector: 'app-dashboard',
@@ -7,10 +9,22 @@ import { Component } from '@angular/core';
   styleUrls: ['./dashboard.component.sass'],
 })
 export class DashboardComponent {
-  constructor(private router: Router) {}
+  constructor(private router: Router, private dialog: MatDialog) {}
 
   logout() {
-    localStorage.clear();
-    this.router.navigate(['../../']);
+    const dialogRef = this.dialog.open(ConfirmModalComponent, {
+      data: {
+        message: `Are you sure you want to LogOut?`,
+      },
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        console.log(result);
+        localStorage.clear();
+        console.log(localStorage);
+        this.router.navigate(['../../']);
+      }
+    });
   }
 }
